+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/catalog/__cmd_group.py
@@ -13,10 +13,9 @@ from azure.cli.core.aaz import *
 
 @register_command_group(
     "devcenter admin catalog",
-    is_preview=True,
 )
 class __CMDGroup(AAZCommandGroup):
-    """Manage catalog with dev center
+    """Manage catalogs.
     """
     pass