+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/image/_show.py
@@ -13,19 +13,18 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "devcenter admin image show",
-    is_preview=True,
 )
 class Show(AAZCommand):
     """Get a gallery image.
 
-    :example: Show
-        az devcenter admin image show --dev-center-name "Contoso" --gallery-name "DefaultDevGallery" --name "{imageName}" --resource-group "rg1"
+    :example: Get
+        az devcenter admin image show --dev-center-name "Contoso" --gallery-name "DefaultDevGallery" --name "ContosoBaseImage" --resource-group "rg1"
     """
 
     _aaz_info = {
-        "version": "2022-11-11-preview",
+        "version": "2023-04-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/galleries/{}/images/{}", "2022-11-11-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/galleries/{}/images/{}", "2023-04-01"],
         ]
     }
 
@@ -47,7 +46,7 @@ class Show(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.dev_center_name = AAZStrArg(
             options=["-d", "--dev-center", "--dev-center-name"],
-            help="The name of the dev center. Use az configure -d dev-center=<dev_center_name> to configure a default.",
+            help="The name of the dev center. Use `az configure -d dev-center=<dev_center_name>` to configure a default.",
             required=True,
             id_part="name",
         )
@@ -142,7 +141,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-11-11-preview",
+                    "api-version", "2023-04-01",
                     required=True,
                 ),
             }
@@ -195,6 +194,9 @@ class Show(AAZCommand):
             properties = cls._schema_on_200.properties
             properties.description = AAZStrType(
                 flags={"read_only": True},
+            )
+            properties.hibernate_support = AAZStrType(
+                serialized_name="hibernateSupport",
             )
             properties.offer = AAZStrType(
                 flags={"read_only": True},