+++ src/devcenter/azext_devcenter/vendored_sdks/devcenter_dataplane/aio/operations/_artifacts_operations.py
@@ -41,7 +41,7 @@ class ArtifactsOperations:
         self._deserialize = deserializer
         self._config = config
 
-    def list_by_environment(
+    def list(
         self,
         environment_name: str,
         user_id: str = "me",
@@ -64,7 +64,7 @@ class ArtifactsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-11-11-preview"
+        api_version = "2023-04-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -74,7 +74,7 @@ class ArtifactsOperations:
 
             if not next_link:
                 # Construct URL
-                url = self.list_by_environment.metadata['url']  # type: ignore
+                url = self.list.metadata['url']  # type: ignore
                 path_format_arguments = {
                     'endpoint': self._serialize.url("self._config.endpoint", self._config.endpoint, 'str', skip_quote=True),
                     'projectName': self._serialize.url("self._config.project_name", self._config.project_name, 'str', max_length=63, min_length=3, pattern=r'^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}$'),
@@ -122,7 +122,7 @@ class ArtifactsOperations:
         return AsyncItemPaged(
             get_next, extract_data
         )
-    list_by_environment.metadata = {'url': '/projects/{projectName}/users/{userId}/environments/{environmentName}/artifacts'}  # type: ignore
+    list.metadata = {'url': '/projects/{projectName}/users/{userId}/environments/{environmentName}/artifacts'}  # type: ignore
 
     def list_by_path(
         self,
@@ -150,7 +150,7 @@ class ArtifactsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-11-11-preview"
+        api_version = "2023-04-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):