+++ src/spring/azext_spring/_build_service.py
@@ -114,7 +114,8 @@ def container_registry_list(cmd, client, resource_group, service):
 
 
 def create_or_update_build(cmd, client, resource_group, service, name=None, builder=None, build_env=None,
-                           build_cpu=None, build_memory=None, source_path=None, artifact_path=None, disable_validation=None):
+                           build_cpu=None, build_memory=None, source_path=None, artifact_path=None,
+                           apms=None, certificates=None, disable_validation=None):
     build_service = BuildService(cmd, client, resource_group, service)
     kwargs = {
         'build_name': name,
@@ -124,6 +125,8 @@ def create_or_update_build(cmd, client, resource_group, service, name=None, buil
         'builder': builder,
         'source_path': source_path,
         'artifact_path': artifact_path,
+        'apms': apms,
+        'certificates': certificates,
         'disable_validation': disable_validation
     }
     build_service.build_and_get_result(4, **kwargs)
@@ -149,14 +152,14 @@ def build_result_list(cmd, client, resource_group, service, build_name=None):
     return client.build_service.list_build_results(resource_group, service, DEFAULT_BUILD_SERVICE_NAME, build_name)
 
 
-def update_build_service(cmd, client, resource_group, service, registry_name=None):
+def update_build_service(cmd, client, resource_group, service, registry_name=None, no_wait=False):
     subscription = get_subscription_id(cmd.cli_ctx)
     service_resource_id = '/subscriptions/{}/resourceGroups/{}/providers/Microsoft.AppPlatform/Spring/{}'.format(subscription, resource_group, service)
     build_service_properties = models.BuildServiceProperties(
         container_registry='{}/containerregistries/{}'.format(service_resource_id, registry_name) if registry_name else None)
     build_service_resource = models.BuildService(
         properties=build_service_properties)
-    return client.build_service.begin_create_or_update(resource_group, service, DEFAULT_BUILD_SERVICE_NAME, build_service_resource)
+    return sdk_no_wait(no_wait, client.build_service.begin_create_or_update, resource_group, service, DEFAULT_BUILD_SERVICE_NAME, build_service_resource)
 
 
 def build_service_show(cmd, client, resource_group, service):